requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Task Queue
celery==5.3.4
redis==5.0.1

# HTTP Client & API Integration
httpx==0.25.2
openai==1.40.0

# File Processing
pillow==10.1.0
python-ffmpeg==2.0.12

# Email (SMTP)
aiosmtplib==3.0.1
jinja2==3.1.2
email-validator==2.1.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Data Validation
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring & Logging
structlog==23.2.0

# File Storage (MinIO)
minio==7.2.0